// src/types/messurementTypes.ts
//ENUM of Food/Drink Messurment Types
export enum MessurementTypes {
  ML = "ml",
  L = "l",
  G = "g",
  KG = "kg",
  PIECES = "pieces",
  CUPS = "cups",
  TABLESPOONS = "tablespoons",
  TEASPOONS = "teaspoons",
  OUNCES = "ounces",
  POUNDS = "pounds",
  MILLIGRAMS = "milligrams",
  GRAMS = "grams",
  KILOGRAMS = "kilograms",
  MILLILITERS = "milliliters",
  LITERS = "liters",
}